divide potrace into own package
